refactor: consolidate 1D B-spline implementation into a unified module and update imports
- Renamed `_bspline_space_1D_impl.py` to `_bspline_space_impl.py` to reflect its broader functionality.
- Merged 1D B-spline space functionality into the new unified implementation module.
- Updated import paths across the codebase to align with the new structure.
- Removed the now redundant `bspline_space_1D.py` file.
