+++ tests/test_bspline.py
@@ -4,8 +4,7 @@ import numpy as np
 import pytest
 
 from pantr.bspline import Bspline
-from pantr.bspline_space import BsplineSpace
-from pantr.bspline_space_1D import BsplineSpace1D
+from pantr.bspline_space import BsplineSpace, BsplineSpace1D
 
 
 class TestBsplineInit:
