+++ tests/test_bspline_space.py
@@ -3,8 +3,7 @@
 import numpy as np
 import pytest
 
-from pantr.bspline_space import BsplineSpace
-from pantr.bspline_space_1D import BsplineSpace1D
+from pantr.bspline_space import BsplineSpace, BsplineSpace1D
 from pantr.quad import PointsLattice
