+++ tests/test_knots_1D.py
@@ -9,11 +9,11 @@ import numpy.testing as nptest
 import numpy.typing as npt
 import pytest
 
-from pantr._bspline_space_1D_impl import (
+from pantr._bspline_space_impl import (
     _get_knots_ends_and_dtype,
     _validate_knot_input,
 )
-from pantr.bspline_space_1D import (
+from pantr.bspline_space import (
     create_cardinal_Bspline_knot_vector,
     create_uniform_open_knot_vector,
     create_uniform_periodic_knot_vector,
